Added support for config values
Added support for config values to the inspector
Pulled common build options out of the Desktop builder and into the
Build.Common.
Set up conditional drawing of the inspector for modules so that (in
this case) per-platform build settings could be hidden when they didn’t
apply to a build.