This version works to make an iOS build on MaxGentlemen
This version works to make an iOS build on MaxGentlemen.  The Facebook
code still needs to be updated to add in the Facebook helper functions
on startup, but it should be pretty close.  I’m checking this in so
that all the core changes are checked in and I can merge in Ben’s
changes.