Bug fixes for the native zip module
Bug fixes for the native zip module
I accidentally swapped the value of the zip name parameter for the
parameter name itself…so that didn’t work well.
This also fixes an issue I was having where it 7zip appeared to append
files instead of creating a new archive if the file already existed.
Also removed some old debug code from iOS that was never running.