Setting up the ability to fail a job and realtime logging in Jenkins with parsable tags
Setting up the ability to fail a job
Any failed AssertTrue or EnsureTrue will result in a job failure.
This will cause any subsequent steps to not run and properly return a
non-zero error code from Unity and the python script if it is being run
from Jenkins.
I’ve updated all the conditionals in the existing code I’m using that
should fail a job to use the assert functions and I’ve gotten a
successful desktop and iOS build using this code.
Also added realtime logging in Jenkins
Added a specific tag to the start of each Igor log line so that we can
easily track down errors.