Got Android up and running for Hats
Got Android up and running for Hats
Finished implementing the Facebook Android Hats module.
Fixed an issue with re-signing the APK that was due to the zip function
including the whole path to each file in the zip file.
Fixed an exception where replacing strings in a file that was not
writable wouldn’t be handled correctly.