batchmode and nographics support as IgorRun.py arguments.
Add argparser arguments for batchmode and nographics, which are just the flags Unity uses with an extra '-' character preceding them. No IgorConfigWindow options for it yet though. Thought about just passing any unconsumed arguments through to Igor, but thought it'd be better to have control over what Unity arguments Igor supports.
Also, have to check UnityEditorInternal.InternalEditorUtility.inBatchMode in the EditorWindow, as attempting to draw anything via GUI when in nographics mode crashes the editor. Would be preferable to move some of the instance-method initialization logic in the window out so that we don't have to create the window at all, but opting to deferr on that problem.