Massive change to move all of Igor into the root Assets directory
Massive change to move all of Igor into the root Assets directory.

Now each module can decide what scripts go in the editor build solution
and what are accessible at runtime.  This should handle migrating
anyone using the original path.  We had to do this in a really terrible
way by renaming the module list file, otherwise the downloader would
pull duplicates of all the files and it would break compilation.

In the Updater script we now have a major upgrade version number so we
can tell the updater that it should only update itself if that version
is different going forward.