Fix -- parsing, build name detection order, remove AppendCommitInfo.
Ragtag edit: Entering -- in a string was breaking the command line parsing. Opted to prevent any string from starting with a space and to disallow '=' in a string so that we can use that as a key for detecting where string params start and end.
Ragtag edit: In IgorBuildDesktop, change the order in which a build name is chosen. If there's an explicit non-config value available, that should take precedence over a last used build name (especially since EditorUserBuildSettings.GetBuildLocation is only set when building through the default Unity build settings dialog).
Remove AppendCommitInfo logic from IgorRun.py, as we don't need to do any of that in the python script. The module can query the system environment variables directly.