Added first pass on Monster Test Framework
Added first pass on Monster Test Framework
This adds the Test list window and the Test State graph editor window
Added two utility modules for the list window and graph editor window
generic code.
The two utilities need a lot of work to make them more generic and trim
out a lot of the unnecessary bits from their previous life.
Moved some utility functions to the RuntimeUtils file.
Added preprocessor blocks around all the runtime code for Igor so we
can conditionally compile it out of runtime and also added conditional
compilation to Monster so it can be disabled separately.
Added the ability to say that a module is needed by another module at
runtime so that it can be used and enabled by a different module when
needed.  This is currently only set up for set scripting defines, but
it can be added to any module.
Fixed the config window so that on save all dependent modules are
actually marked as enabled and not just shown as enabled in the window.
Fixed some typos in the IgorRun.py script and made the default behavior
to not update in Unity (this was breaking some deployments because of
the last huge change)
Set up the appropriate framework for Igor to be able to run modules at
runtime within a built game (this is currently only used by Monster,
but can be used in other ways.  Look at the Monster source for more
info)
Set up a temp value for drawing an inspector that’s dependent on an
environment variable so it would have some value and allow the module
to be enabled when previewing job steps in the inspector.
Added un-deploy for BitTorrent Sync so that files can be copied from
the Sync folder in addition to being copied to them.
Added an un-package step and implemented it for Zip so that we can
unzip built executables as a job step before running Monster tests.

Monster is currently barely born.  It has the ability to launch an
executable, but there isn’t any additional testing functionality built
in.  This is the starting point and it will get a lot of cool new
functionality soon, but I needed to checkpoint all this code so I could
start cleaning it up a bit more.  I also really badly need to do a
documentation pass on the two new Utilities and Monster itself.  This
will be coming shortly.