Adding in a few new cases of when we need to kill Unity during a build job
Adding in a few new cases of when we need to kill Unity during a build
job.

These haven’t happened more than once, so this code is in place with
the hope that it works, but I haven’t really tested it yet.  So far
with successful builds (and even other build failures) this code hasn’t
killed the process unnecessarily, so at worst it should just fail to
kill the process under these conditions.