Workaround for a bad FTP server that we're using
Workaround for a bad FTP server that we're using

In our situation the file will completely upload to the server and then
the client attempts a read (presumably to verify the file is uploaded)
and times out while reading the response back.  The file has already
been successfully uploaded, but it still throws an exception and stops
the rest of the build.  For now I’m just ignoring those exceptions, but
we may want to make this configurable at some point.