Fix for the delayed compilation error handling
Fix for the delayed compilation error handling

The asset database refresh is too early in the process so a failure in
a different compilation pass would have caused the same early out issue
that we were trying to fix.  Loading the Editor module into memory is
the sign that Unity has finished its first attempt at compiling all the
modules (but it still does another compile if the first attempt fails,
so we can still catch error messages after the module is loaded).