Another bad startup scenario check
Another bad startup scenario check.  Somehow there’s a new scenario
where we get to the window layout dialog before we get the loading
editor assembly dll into memory.  This is an attempt to see if the log
message is being logged properly in a way that we can use it.  It
didn’t look like it got logged before user interaction (which would
make it impossible for us to check for this from the log alone), so we
may need to set up some sort of timer to make sure that compilation and
import doesn’t stop logging for more than 10 minutes or something
reasonably long to make sure that we catch any dialogs that might pop
up even though we passed buildmachine :-(